Tried to make it work with doubles, now broken

+++ HW4/LinkedList.c
@@ -64,7 +64,7 @@ int LListLength(LinkedListPtr_t L){
  *
  * @returns	the data from the selected node as a double float
  */
-long getNodeDataInLList(LinkedListPtr_t L, int pos) {
+double getNodeDataInLList(LinkedListPtr_t L, int pos) {
     ListNodePtr_t p = L->head;
     int k = 1;
